+++ tests/window_test.c
@@ -162,7 +162,6 @@ bool windowTest()
     PalResult result;
     PalWindow* window = nullptr;
     PalWindowCreateInfo createInfo = {0};
-    PalVideoFeatures2 features;
     bool running = false;
 
     // event driver
@@ -194,7 +193,7 @@ bool windowTest()
     }
 
     // get video system features
-    features = palGetVideoFeaturesEx();
+    PalVideoFeatures64 features = palGetVideoFeaturesEx();
 
     // fill the create info struct
     createInfo.monitor = nullptr; // use default monitor
@@ -204,7 +203,7 @@ bool windowTest()
     createInfo.style = PAL_WINDOW_STYLE_RESIZABLE;
 
     // check if we support decorated windows (title bar, close etc)
-    if (!(features & PAL_VIDEO_FEATURE_DECORATED_WINDOW)) {
+    if (!(features & PAL_VIDEO_FEATURE64_DECORATED_WINDOW)) {
         // if we dont support, we need to create a borderless window
         // and create the decorations ourselves
         createInfo.style |= PAL_WINDOW_STYLE_BORDERLESS;